Refactor web3
