+++ Dashboard/web3/newHeads.ts
@@ -1,0 +1,6 @@
+import { BlockHeader } from "web3-eth";
+import web3 from './web3Provider';
+
+
+
+export const Subscription = web3.eth.subscribe('newBlockHeaders');
